+++ Struktury/Funkcje/do_segments_intersect.cpp
@@ -3,10 +3,10 @@
 // Sprawdzenie czy odcinki p1p2 p3p4 się przeciniają
 bool do_segments_intersect(Point p1, Point p2, Point p3, Point p4)
 {
-    float d1 = det(p1,p2,p3);
-    float d2 = det(p1,p2,p4);
-    float d3 = det(p3,p4,p1);
-    float d4 = det(p3,p4,p2);
+    double d1 = det(p1,p2,p3);
+    double d2 = det(p1,p2,p4);
+    double d3 = det(p3,p4,p1);
+    double d4 = det(p3,p4,p2);
 
     // Przecinają się gdy:
     // WARUNEK 2: każdy odcinek przecina prostą wyznaczoną przez drugi odcinek
