+++ Struktury/Point.cpp
@@ -1,6 +1,6 @@
 #include "Point.h"
 
-Point::Point(float x, float y) : x(x), y(y) {}
+Point::Point(double x, double y) : x(x), y(y) {}
 
 // Operator mniejszosci
 bool Point::operator<(const Point& other) const {
